test: share serialized config bundle through fixture
